Refactored fetch_as_dict function to be labeled as a util
